refactor: add updateProfileRatings and calculatePostUpdates functions
