+++ app/src/main/java/com/github/lookupgroup27/lookup/ui/feed/Feed.kt
@@ -1,4 +1,3 @@
-// File: com/github/lookupgroup27/lookup/ui/feed/FeedScreen.kt
 package com.github.lookupgroup27.lookup.ui.feed
 
 import android.Manifest
@@ -27,9 +26,8 @@ import com.github.lookupgroup27.lookup.R
 import com.github.lookupgroup27.lookup.model.feed.ProximityAndTimePostFetcher
 import com.github.lookupgroup27.lookup.model.location.LocationProviderSingleton
 import com.github.lookupgroup27.lookup.model.post.Post
+import com.github.lookupgroup27.lookup.model.profile.UserProfile
 import com.github.lookupgroup27.lookup.ui.feed.components.PostItem
-import com.github.lookupgroup27.lookup.ui.feed.components.calculatePostUpdates
-import com.github.lookupgroup27.lookup.ui.feed.components.updateProfileRatings
 import com.github.lookupgroup27.lookup.ui.navigation.BottomNavigationMenu
 import com.github.lookupgroup27.lookup.ui.navigation.LIST_TOP_LEVEL_DESTINATION
 import com.github.lookupgroup27.lookup.ui.navigation.NavigationActions
@@ -63,9 +61,7 @@ fun FeedScreen(
     profileViewModel: ProfileViewModel,
     initialNearbyPosts: List<Post>? = null
 ) {
-  // Fetch user profile
-  LaunchedEffect(Unit) { profileViewModel.fetchUserProfile() }
-
+  profileViewModel.fetchUserProfile()
   val profile by profileViewModel.userProfile.collectAsState()
   val user = FirebaseAuth.getInstance().currentUser
   val isUserLoggedIn = user != null
@@ -115,6 +111,48 @@ fun FeedScreen(
         postRatings[post.uid] = initialRating.toMutableList()
       }
     }
+  }
+
+  // Helper function to update user profile ratings
+  fun updateProfileRatings(
+      currentProfile: UserProfile?,
+      postUid: String,
+      starsCount: Int,
+      username: String,
+      bio: String,
+      email: String
+  ): UserProfile {
+    val updatedRatings =
+        currentProfile?.ratings?.toMutableMap()?.apply { this[postUid] = starsCount }
+            ?: mutableMapOf(postUid to starsCount)
+
+    return currentProfile?.copy(
+        username = username, bio = bio, email = email, ratings = updatedRatings)
+        ?: UserProfile(username = username, bio = bio, email = email, ratings = updatedRatings)
+  }
+
+  // Helper function to calculate post updates
+  fun calculatePostUpdates(
+      post: Post,
+      userEmail: String,
+      starsCount: Int,
+      oldStarCounts: Int
+  ): Post {
+    val isReturningUser = post.ratedBy.contains(userEmail)
+    val newStarsCount =
+        if (isReturningUser) {
+          post.starsCount - oldStarCounts + starsCount
+        } else {
+          post.starsCount + starsCount
+        }
+    val newUsersNumber = if (isReturningUser) post.usersNumber else post.usersNumber + 1
+    val newAvg = newStarsCount.toDouble() / newUsersNumber
+
+    return post.copy(
+        averageStars = newAvg,
+        starsCount = newStarsCount,
+        usersNumber = newUsersNumber,
+        ratedBy = if (!isReturningUser) post.ratedBy + userEmail else post.ratedBy)
   }
 
   // Background Box with gradient overlay using drawBehind for efficiency.
@@ -188,32 +226,24 @@ fun FeedScreen(
                               PostItem(
                                   post = post,
                                   starStates =
-                                      postRatings[post.uid] ?: List(NUMBER_OF_STARS) { false },
+                                      postRatings[post.uid] ?: mutableListOf(false, false, false),
                                   onRatingChanged = { newRating ->
                                     val oldPostRatings =
-                                        postRatings[post.uid] ?: List(NUMBER_OF_STARS) { false }
+                                        postRatings[post.uid] ?: mutableListOf(false, false, false)
                                     val oldStarCounts = oldPostRatings.count { it }
                                     postRatings[post.uid] = newRating.toList()
                                     val starsCount = newRating.count { it }
 
-                                    // Update user profile ratings using the utility function
+                                    // Update user profile ratings
                                     val newProfile =
                                         updateProfileRatings(
-                                            currentProfile = profile,
-                                            postUid = post.uid,
-                                            starsCount = starsCount,
-                                            username = username,
-                                            bio = bio,
-                                            email = email)
+                                            profile, post.uid, starsCount, username, bio, email)
                                     profileViewModel.updateUserProfile(newProfile)
 
-                                    // Update post details using the utility function
+                                    // Update post details
                                     val updatedPost =
                                         calculatePostUpdates(
-                                            post = post,
-                                            userEmail = userEmail,
-                                            starsCount = starsCount,
-                                            oldStarCounts = oldStarCounts)
+                                            post, userEmail, starsCount, oldStarCounts)
                                     postsViewModel.updatePost(updatedPost)
                                   })
                             }
